+++ lib/snowflex.ex
@@ -92,6 +92,7 @@ defmodule Snowflex do
         data =
           row
           |> elem(index)
+          |> handle_encoding()
           |> to_string_if_charlist()
           |> map_null_to_nil(map_nulls_to_nil?)
 
@@ -107,6 +108,17 @@ defmodule Snowflex do
 
   defp map_null_to_nil(:null, true), do: nil
   defp map_null_to_nil(data, _), do: data
+
+  defp handle_encoding(data) when is_list(data) do
+    raw = :erlang.list_to_binary(data)
+
+    case :unicode.characters_to_binary(raw) do
+      utf8 when is_binary(utf8) -> utf8
+      _ -> :unicode.characters_to_binary(raw, :latin1)
+    end
+  end
+
+  defp handle_encoding(data), do: data
 
   defp cast_row(row, schema) do
     schema