+++ test/snowflex/connection_test.exs
@@ -21,6 +21,10 @@ defmodule Snowflex.ConnectionTest do
       {:ok, %{}}
     end
 
+    def handle_call({:sql_query, "return_utf_16"}, _from, state) do
+      {:reply, {:ok, {:selected, ['col'], [{'CVS PharmacyÂ®'}]}}, state}
+    end
+
     def handle_call({:sql_query, "insert " <> _}, _from, state) do
       {:reply, {:ok, {:updated, 123}}, state}
     end
@@ -49,6 +53,13 @@ defmodule Snowflex.ConnectionTest do
       start_supervised!(SnowflakeConnection)
 
       assert {:updated, 123} == SnowflakeConnection.execute("insert query")
+    end
+
+    test "should execute a sql query with utf-16 artifacts and scrub it" do
+      start_supervised!(SnowflakeConnection)
+      # above we set this to return 'CVS PharmacyÂ®', as we saw in the real world, and then
+      # here we ensure that we just have the ® character
+      assert [%{"col" => "CVS Pharmacy®"}] == SnowflakeConnection.execute("return_utf_16")
     end
   end